requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
flask_pymongo
dnspython
pymongo[srv]
email_validator
# ecdsa
# paramiko
# pycrypto
